feat(unique-names-generator): Prepare for version 4
* Folder restructure
* Add mandatory `dictionaries` options
* Enable tree-shaking

BREAKING CHANGE: New mandatory `dictionaries` option
