--- index.ts
@@ -1,2 +1,0 @@
-export * from './lib/index';
-export * from './lib/dictionaries/index';
