+++ src/dictionaries/adjectives.ts
