+++ src/dictionaries/animals.ts
