+++ src/dictionaries/colors.ts
