+++ src/dictionaries/index.ts
