+++ src/index.ts
@@ -1,0 +1,3 @@
+export * from './unique-names-generator';
+export * from './unique-names-generator.constructor';
+export * from './dictionaries/index';
